+++ socialuni-server/socialuni-common-api/src/main/java/com/socialuni/social/common/api/config/SocialRequestUserConfig.java
@@ -4,5 +4,5 @@ public interface SocialRequestUserConfig {
     //支持用户重写token，有那种根据token，获取其他token返回的情况
     String getToken();
 
-    Integer getUserId();
+    Long getUserId();
 }
