+++ socialuni-server/socialuni-community-sdk/src/main/java/com/socialuni/social/community/sdk/dao/DO/SocialuniTalkImgDO.java
@@ -24,5 +24,5 @@ import java.io.Serializable;
 @EqualsAndHashCode(callSuper = true)
 public class SocialuniTalkImgDO extends SocialuniImgBaseDo implements Serializable {
     @Column(nullable = false, updatable = false)
-    private Integer talkId;
+    private Long talkId;
 }
